@@ -1,12 +1,11 @@
-import Vue from "vue";
-import App from "./App.vue";
-import router from "./router";
-import store from "./store";
+import { createApp } from 'vue'
+import App from './App.vue'
+import router from './router'
+import store from './store'
+import OutlinePoint from './components/OutlinePoint.vue'
 
-Vue.config.productionTip = false;
-
-new Vue({
-  router,
-  store,
-  render: h => h(App)
-}).$mount("#app");
+createApp(App)
+  .use(store)
+  .use(router)
+  .component('outline-point', OutlinePoint)
+  .mount('#app')
